refactor build_body method
NOTE: this was originally submitted as a PR to the Ginja/puppet_stack develop branch.

This method was choking on Unicode. Has something in the Ruby core String#gsub method.

In any case, there's much easier, more consistent way to parse this data...

Rather than trying to skim fat off the top of the String, load the 'puppet' gem so that the YAML parser understands what it is when it encounters a Puppet data object. Then, simply access the facts Hash on that object via the Object#values method.